% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AverageAge.R
\name{Average}
\alias{Average}
\title{Average of faculty ages}
\usage{
Average()
}
\value{
Returns the average of \code{actualData}.
}
\description{
\code{Average} returns the average of the data set actualData.
}

