% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AverageAge.R
\name{Max}
\alias{Max}
\title{Maximum value of faculty ages}
\usage{
Max()
}
\value{
Returns the maximum of \code{actualData}.
}
\description{
\code{Max} returns the maximum value of the data set actualData.
}

