% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AverageAge.R
\name{Min}
\alias{Min}
\title{Minimum value of faculty ages}
\usage{
Min()
}
\value{
Returns the minimum of \code{actualData}.
}
\description{
\code{Min} returns the minimum value of the data set actualData.
}

