% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AverageAge.R
\name{PlotHist}
\alias{PlotHist}
\title{Histogram  of faculty ages}
\usage{
PlotHist()
}
\value{
Returns the histogram of \code{actualData}.
}
\description{
\code{AgeHist} returns the histogram of the data set actualData.
}

