% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AverageAge.R
\name{PrintAll}
\alias{PrintAll}
\title{Information  of faculty ages}
\usage{
PrintAll()
}
\value{
Returns the information of \code{actualData}.
}
\description{
\code{AgeHist} returns all the above information of actualData.
}

