% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AverageAge.R
\name{RequiredFunction}
\alias{RequiredFunction}
\title{Data extrapolation}
\usage{
RequiredFunction()
}
\value{
Returns the average of \code{actualData}.
}
\description{
\code{RequiredFunction} returns the data set actualData.
}

